requirements.txt
asgiref==3.7.2
Django==5.0.2
django-phonenumber-field==7.3.0
phonenumbers==8.13.29
sqlparse==0.4.4
tzdata==2023.4
